adding Clarity to problem 5
When reading through this problem, it was really easy to get confused as to what the end goal was. As it was, it would be logical to think that tax was the amount of tax  in pennies that was needed to add rather than the percentage of tax that would be needed to add.